% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{projectMolGrade}
\alias{projectMolGrade}
\title{projectMolGrade}
\usage{
projectMolGrade(newexp, geneSymbols)
}
\arguments{
\item{newexp}{gene expression matrix or dataframe with gene in row and sample in columns.}

\item{geneSymbols}{vector of gene symbols for the newexp dataset (simply set to rownames if the nexzexp is already in single values per gene symbols)}
}
\value{
data frame of three oroejctions
}
\description{
projectMolGrade
}
